Make functions more atomic, add color and logging.

+++ bin/uninstall_openshift.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 
 script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-. $script_dir/utils/bootstrap.sh
+source $script_dir/utils/bootstrap.sh
 
+# Reset script_dir, which is apparently overridden during bootstrap
+script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 $script_dir/destroy_openshift.sh
 
 cd "$build_path"
-rm -rf openshift
+if [ -d openshift ]; then
+  rm -rf openshift
+fi
